moved authentication trait to base sync project
